deepseek-OCR/serverless/requirements_serverless.txt
# DeepSeek-OCR RunPod Serverless Dependencies
# For use with: runpod/pytorch:2.2.0-py3.10-cuda12.1.1-devel-ubuntu22.04

# ML Core (PyTorch comes from base image)
transformers==4.46.3
tokenizers==0.20.3
accelerate>=0.26.0
einops
easydict
addict

# Image Processing
Pillow>=10.0.0
numpy
PyMuPDF

# HTTP Client
requests>=2.31.0
